Reorganised dependencies
