+++ src/solve.jl
@@ -57,271 +57,15 @@ macro withprogress(progress, exprs...)
         $(exprs[end])
       end
     end |> esc
-  end
-
-function __solve(prob::OptimizationProblem, opt, data = DEFAULT_DATA;
-                 maxiters::Number = 0, cb = (args...) -> (false),
-                 progress = false, save_best = true, kwargs...)
-
-    if data != DEFAULT_DATA
-      maxiters = length(data)
-    else
-      if maxiters <= 0.0
-        error("The number of maxiters has to be a non-negative and non-zero number.")
-      end
-      data = take(data, maxiters)
-    end
-
-    # Flux is silly and doesn't have an abstract type on its optimizers, so assume
-    # this is a Flux optimizer
-    θ = copy(prob.u0)
-    ps = Flux.params(θ)
-
-    t0 = time()
-
-    local x, min_err, _loss
-    min_err = typemax(eltype(prob.u0)) #dummy variables
-    min_opt = 1
-
-    f = instantiate_function(prob.f,prob.u0,prob.f.adtype,prob.p)
-
-    @withprogress progress name="Training" begin
-      for (i,d) in enumerate(data)
-        gs = Flux.Zygote.gradient(ps) do
-            x = prob.f(θ,prob.p, d...)
-            first(x)
-          end
-        x = f.f(θ, prob.p, d...)
-        cb_call = cb(θ, x...)
-        if !(typeof(cb_call) <: Bool)
-          error("The callback should return a boolean `halt` for whether to stop the optimization process. Please see the sciml_train documentation for information.")
-        elseif cb_call
-          break
-        end
-        msg = @sprintf("loss: %.3g", x[1])
-        progress && ProgressLogging.@logprogress msg i/maxiters
-        Flux.update!(opt, ps, gs)
-
-        if save_best
-          if first(x) < first(min_err)  #found a better solution
-            min_opt = opt
-            min_err = x
-          end
-          if i == maxiters  #Last iteration, revert to best.
-            opt = min_opt
-            cb(θ,min_err...)
-          end
-        end
-      end
-    end
-
-    _time = time()
-
-    SciMLBase.build_solution(prob, opt, θ, x[1])
-    # here should be build_solution to create the output message
 end
 
-
-decompose_trace(trace::Optim.OptimizationTrace) = last(trace)
 decompose_trace(trace) = trace
 
-function __solve(prob::OptimizationProblem, opt::Optim.AbstractOptimizer,
-                 data = DEFAULT_DATA;
-                 maxiters = nothing,
-                 cb = (args...) -> (false),
-                 progress = false,
-                 kwargs...)
-    local x, cur, state
-
-    if data != DEFAULT_DATA
-        maxiters = length(data)
-    end
-
-    cur, state = iterate(data)
-
-    function _cb(trace)
-        cb_call = opt == NelderMead() ? cb(decompose_trace(trace).metadata["centroid"],x...) : cb(decompose_trace(trace).metadata["x"],x...)
-        if !(typeof(cb_call) <: Bool)
-            error("The callback should return a boolean `halt` for whether to stop the optimization process.")
-        end
-        cur, state = iterate(data, state)
-        cb_call
-    end
-
-    if !(isnothing(maxiters)) && maxiters <= 0.0
-        error("The number of maxiters has to be a non-negative and non-zero number.")
-    elseif !(isnothing(maxiters))
-        maxiters = convert(Int, maxiters)
-    end
-
-    f = instantiate_function(prob.f,prob.u0,prob.f.adtype,prob.p)
-
-    !(opt isa Optim.ZerothOrderOptimizer) && f.grad === nothing && error("Use OptimizationFunction to pass the derivatives or automatically generate them with one of the autodiff backends")
-
-    _loss = function(θ)
-        x = f.f(θ, prob.p, cur...)
-        return first(x)
-    end
-
-    fg! = function (G,θ)
-        if G !== nothing
-            f.grad(G, θ, cur...)
-        end
-        return _loss(θ)
-    end
-
-    if opt isa Optim.KrylovTrustRegion
-        optim_f = Optim.TwiceDifferentiableHV(_loss, fg!, (H,θ,v) -> f.hv(H,θ,v,cur...), prob.u0)
-    else
-        optim_f = TwiceDifferentiable(_loss, (G, θ) -> f.grad(G, θ, cur...), fg!, (H,θ) -> f.hess(H,θ,cur...), prob.u0)
-    end
-
-    original = Optim.optimize(optim_f, prob.u0, opt,
-                              !(isnothing(maxiters)) ?
-                                Optim.Options(;extended_trace = true,
-                                               callback = _cb,
-                                               iterations = maxiters,
-                                               kwargs...) :
-                                Optim.Options(;extended_trace = true,
-                                               callback = _cb, kwargs...))
-    SciMLBase.build_solution(prob, opt, original.minimizer,
-                             original.minimum; original=original)
-end
-
-function __solve(prob::OptimizationProblem, opt::Union{Optim.Fminbox,Optim.SAMIN},
-                 data = DEFAULT_DATA;
-                 maxiters = nothing,
-                 cb = (args...) -> (false),
-                 progress = false,
-                 kwargs...)
-
-    local x, cur, state
-
-    if data != DEFAULT_DATA
-        maxiters = length(data)
-    end
-
-    cur, state = iterate(data)
-
-    function _cb(trace)
-        cb_call = !(opt isa Optim.SAMIN) && opt.method == NelderMead() ? cb(decompose_trace(trace).metadata["centroid"],x...) : cb(decompose_trace(trace).metadata["x"],x...)
-        if !(typeof(cb_call) <: Bool)
-            error("The callback should return a boolean `halt` for whether to stop the optimization process.")
-        end
-        cur, state = iterate(data, state)
-        cb_call
-    end
-
-    if !(isnothing(maxiters)) && maxiters <= 0.0
-        error("The number of maxiters has to be a non-negative and non-zero number.")
-    elseif !(isnothing(maxiters))
-        maxiters = convert(Int, maxiters)
-    end
-
-    f = instantiate_function(prob.f,prob.u0,prob.f.adtype,prob.p)
-
-    !(opt isa Optim.ZerothOrderOptimizer) && f.grad === nothing && error("Use OptimizationFunction to pass the derivatives or automatically generate them with one of the autodiff backends")
-
-    _loss = function(θ)
-        x = f.f(θ, prob.p, cur...)
-        return first(x)
-    end
-    fg! = function (G,θ)
-        if G !== nothing
-            f.grad(G, θ, cur...)
-        end
-
-        return _loss(θ)
-    end
-    optim_f = OnceDifferentiable(_loss, (G, θ) -> f.grad(G, θ, cur...), fg!, prob.u0)
-
-    original = Optim.optimize(optim_f, prob.lb, prob.ub, prob.u0, opt,
-                              !(isnothing(maxiters)) ? Optim.Options(;
-                              extended_trace = true, callback = _cb,
-                              iterations = maxiters, kwargs...) :
-                              Optim.Options(;extended_trace = true,
-                              callback = _cb, kwargs...))
-    SciMLBase.build_solution(prob, opt, original.minimizer,
-                             original.minimum; original=original)
-end
-
-
-function __solve(prob::OptimizationProblem, opt::Optim.ConstrainedOptimizer,
-                 data = DEFAULT_DATA;
-                 maxiters = nothing,
-                 cb = (args...) -> (false),
-                 progress = false,
-                 kwargs...)
-
-    local x, cur, state
-
-    if data != DEFAULT_DATA
-        maxiters = length(data)
-    end
-
-    cur, state = iterate(data)
-
-      function _cb(trace)
-      cb_call = cb(decompose_trace(trace).metadata["x"],x...)
-      if !(typeof(cb_call) <: Bool)
-          error("The callback should return a boolean `halt` for whether to stop the optimization process.")
-      end
-      cur, state = iterate(data, state)
-      cb_call
-    end
-
-    if !(isnothing(maxiters)) && maxiters <= 0.0
-        error("The number of maxiters has to be a non-negative and non-zero number.")
-    elseif !(isnothing(maxiters))
-        maxiters = convert(Int, maxiters)
-    end
-
-    f = instantiate_function(prob.f,prob.u0,prob.f.adtype,prob.p,prob.ucons === nothing ? 0 : length(prob.ucons))
-
-    f.cons_j ===nothing && error("Use OptimizationFunction to pass the derivatives or automatically generate them with one of the autodiff backends")
-
-    _loss = function(θ)
-        x = f.f(θ, prob.p, cur...)
-        return x[1]
-    end
-    fg! = function (G,θ)
-        if G !== nothing
-            f.grad(G, θ, cur...)
-        end
-        return _loss(θ)
-    end
-    optim_f = TwiceDifferentiable(_loss, (G, θ) -> f.grad(G, θ, cur...), fg!, (H,θ) -> f.hess(H, θ, cur...), prob.u0)
-
-    cons! = (res, θ) -> res .= f.cons(θ);
-
-    cons_j! = function(J, x)
-        f.cons_j(J, x)
-    end
-
-    cons_hl! = function (h, θ, λ)
-        res = [similar(h) for i in 1:length(λ)]
-        f.cons_h(res, θ)
-        for i in 1:length(λ)
-            h .+= λ[i]*res[i]
-        end
-    end
-
-    lb = prob.lb === nothing ? [] : prob.lb
-    ub = prob.ub === nothing ? [] : prob.ub
-    optim_fc = TwiceDifferentiableConstraints(cons!, cons_j!, cons_hl!, lb, ub, prob.lcons, prob.ucons)
-
-    original = Optim.optimize(optim_f, optim_fc, prob.u0, opt,
-                              !(isnothing(maxiters)) ? Optim.Options(;
-                                extended_trace = true, callback = _cb,
-                                iterations = maxiters, kwargs...) :
-                                Optim.Options(;extended_trace = true,
-                                callback = _cb, kwargs...))
-    SciMLBase.build_solution(prob, opt, original.minimizer,
-                             original.minimum; original=original)
-end
-
-
 function __init__()
+    @require Flux="587475ba-b771-5e3f-ad9e-33799f191a9c" include("solve_flux.jl")
+
+    @require Optim="429524aa-4258-5aef-a3af-852621145aeb" include("solve_optim.jl")
+
     @require BlackBoxOptim="a134a8b2-14d6-55f6-9291-3336d3ab0209" begin
         decompose_trace(opt::BlackBoxOptim.OptRunController) = BlackBoxOptim.best_candidate(opt)
